+++ test/lucene/monitor_test.clj
@@ -186,11 +186,11 @@
 
   (testing "default matching mode"
     (with-open [monitor (m/monitor {:default-match-mode :highlight} [{:id "1" :query "test"}])]
-      (is (= [{:id "1"
+      (is (= [{:id         "1"
                :highlights {"text" [{:start-position 1
-                                     :end-position 1
-                                     :start-offset 7
-                                     :end-offset 11}]}}]
+                                     :end-position   1
+                                     :start-offset   7
+                                     :end-offset     11}]}}]
              (m/match-string monitor "prefix test suffix")))))