+++ CHANGELOG.md
@@ -4,6 +4,8 @@ For a list of breaking changes, check [here](#breaking-changes).
 
 ## Unreleased
 
+- Replace `Jsonista` with `charred` 
+
 ## 1.0.2 (2023-08-31)
 
 - initial release